Split the single module file into submodules
